# Conflicts:
#	folib-web-core/src/main/java/com/veadan/folib/components/syncartifact/RawSyncArtifactProvider.java

+++ folib-storage/folib-storage-api/src/main/java/com/veadan/folib/providers/io/StorageFileSystemProvider.java
@@ -13,6 +13,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.util.FileSystemUtils;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.URI;
 import java.nio.channels.FileChannel;
@@ -117,6 +118,11 @@ public abstract class StorageFileSystemProvider
         };
     }
 
+    @Override
+    public InputStream newInputStream(Path path, OpenOption... options) throws IOException {
+        return target.newInputStream(unwrap(path), options);
+    }
+
     protected Iterator<Path> createRepositoryDsIterator(LayoutFileSystem rfs,
                                                         Iterator<Path> iterator) {
 
@@ -168,10 +174,10 @@ public abstract class StorageFileSystemProvider
         }
 
         RepositoryPath repositoryPath = (RepositoryPath) path;
-        if (!Files.exists(repositoryPath.getTarget())) {
-            throw new NoSuchFileException(unwrap(repositoryPath).toString());
-        }
-
+//        if (!Files.exists(repositoryPath.getTarget())) {
+//            throw new NoSuchFileException(unwrap(repositoryPath).toString());
+//
+//        }
         if (!Files.isDirectory(repositoryPath)) {
             doDeletePath(repositoryPath, force, true);
 
@@ -276,7 +282,9 @@ public abstract class StorageFileSystemProvider
         }
 
         RepositoryPath trashPath = getTrashPath(repositoryPath);
-
+        if (!Files.exists(repositoryPath)) {
+            return;
+        }
         try {
             Files.move(repositoryPath.getTarget(),
                     trashPath.getTarget(),
